Update feat #4 from Mobile

See merge request Elayve/kookie!38

+++ api/src/main/java/com/api/kookie/controllers/profile/ProfileController.java
@@ -36,7 +36,7 @@ public class ProfileController {
     }
 
     @GetMapping("/favorite/add")
-    public ResponseEntity<Boolean> addFavorite(@RequestParam Integer profileId, @RequestParam Integer recetteId) {
+    public ResponseEntity<Boolean> addFavorite(@RequestParam Integer profileId, @RequestParam Long recetteId) {
         LOGGER.debug("[ProfileController, addFavorite] profileId : " + profileId + " recetteId : " + recetteId);
 
         if (profileId == null || recetteId == null) {
